# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Remote Chrome DevTools Protocol",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:12345",
    "basePath": "/",
    "paths": {
...
        "/{session_code}/attribute": {
            "post": {
                "description": "Get or Set element attribute",
                "tags": [
                    "Attributes"
                ],
                "summary": "Get or Set element attribute",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.attrRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.attrResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/close": {
            "delete": {
                "description": "Close browser",
                "tags": [
                    "Close browser"
                ],
                "summary": "Close browser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/inner_html": {
            "post": {
                "description": "Get inner HTML from element",
                "tags": [
                    "Inner HTML"
                ],
                "summary": "Get inner HTML from element",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.innerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.innerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/inner_text": {
            "post": {
                "description": "Get inner Text from element",
                "tags": [
                    "Inner Text"
                ],
                "summary": "Get inner Text from element",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.innerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.innerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/javascript": {
            "post": {
                "description": "Execute script",
                "tags": [
                    "Javascript"
                ],
                "summary": "Execute script",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.jsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.jsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
        "/{session_code}/mouse": {
            "post": {
                "description": "Mouse Event",
                "tags": [
                    "Mouse Event"
                ],
                "summary": "Mouse Event",
                "parameters": [
                    {
                        "description": "content data",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.mouseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.dataResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.XY": {
            "type": "object",
            "properties": {
                "x": {
                    "description": "X coordinate",
                    "type": "number",
                    "format": "float64",
                    "example": 100
                },
                "y": {
                    "type": "number",
                    "format": "float64",
                    "example": 100
                }
            }
        },
        "main.attrRequest": {
            "type": "object",
            "properties": {
                "attribute": {
                    "description": "Attribute name",
                    "type": "string",
                    "format": "string",
                    "example": "name"
                },
                "data": {
                    "description": "Content data",
                    "type": "string",
                    "format": "string",
                    "example": "el_name"
                },
                "selector": {
                    "description": "XPATH Selector",
                    "type": "string",
                    "format": "string",
                    "example": "id(\"btn_login\")"
                },
                "set": {
                    "description": "Setter: if true set value to attribute, otherwise retrieve data from attribute",
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
        "main.attrResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "evidence": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "main.dataResponse": {
            "type": "object",
            "properties": {
                "document": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "evidence": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "main.innerRequest": {
            "type": "object",
            "properties": {
                "selector": {
                    "description": "XPATH Selector",
                    "type": "string",
                    "format": "string",
                    "example": "id(\"btn_login\")"
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                }
            }
        },
        "main.innerResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "evidence": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "main.internalError": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "inner_error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.jsRequest": {
            "type": "object",
            "properties": {
                "script": {
                    "description": "javascript to be executed",
                    "type": "string",
                    "example": ""
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
        "main.jsResponse": {
            "type": "object",
            "properties": {
                "content": {},
                "document": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/main.internalError"
                },
                "evidence": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "main.mouseRequest": {
            "type": "object",
            "properties": {
                "button": {
                    "description": "Mouse Button:\n* left - Left Button\n* middle - Wheel Button\n* right - Right Button",
                    "type": "string",
                    "x-nullable": true
                },
                "event": {
                    "description": "Event type:\n* click - Click Event\n* mouseover - Mouseover Event\n* mouseout - Mouseout Event\n* scroll - Scroll Event\n* mousedown - Mousedown Event\n* mouseup - Mouseup Event",
                    "type": "string",
                    "format": "string",
                    "example": "click"
                },
                "position": {
                    "description": "Click at position",
                    "$ref": "#/definitions/main.XY"
                },
                "selector": {
                    "description": "XPATH Selector",
                    "type": "string",
                    "format": "string",
                    "example": "id(\"btn_login\")"
                },
                "timeout": {
                    "description": "Timeout in nanoseconds",
                    "type": "integer",
                    "format": "int64",
                    "example": 1000000000
                },
                "times": {
                    "description": "Number of consecutive clicks",
                    "type": "integer",
                    "format": "int",
                    "example": 1
                },
                "wait_for": {
                    "description": "Wait for XPATH visible",
                    "type": "string",
                    "format": "string",
                    "example": "//form/input[2]"
                }
            }
        },
...
    }
}